+++ frontend/components/timetable-date.vue
@@ -1,26 +1,26 @@
 <!--日付表示component-->
 <template>
   <!--祝日判断-->
-  <template v-if="!props.isHoliday">
+  <template v-if="props.isHoliday">
+    <!--祝日-->
+    <th class="horizontal-writing date-holiday-cell">
+      <p class="font-size-l">{{ props.date }}</p>
+      <p :class="[fontSizeHolodayTitle()]">{{ props.holidayTitle }}</p>
+    </th>
+  </template>
+  <template v-else>
     <!--日曜判断-->
     <template v-if="props.dayOfWeek === 0">
-      <th class="date-holiday-cell">
-        <p class="font-size-l date-cell-margin">{{ props.date }}</p>
+      <th class="horizontal-writing date-holiday-cell">
+        <p class="font-size-l">{{ props.date }}</p>
       </th>
     </template>
     <!--日曜以外-->
     <template v-else>
-      <th class="date-cell">
-        <p class="font-size-l date-cell-margin">{{ props.date }}</p>
+      <th class="horizontal-writing date-cell">
+        <p class="font-size-l">{{ props.date }}</p>
       </th>
     </template>
-  </template>
-  <!--祝日-->
-  <template v-else>
-    <th class="date-holiday-cell">
-      <p class="font-size-l date-cell-margin">{{ props.date }}</p>
-      <p class="date-cell-margin" :class="[fontSizeHolodayTitle()]">{{ props.holidayTitle }}</p>
-    </th>
   </template>
 </template>
 <script lang="ts" setup>
@@ -44,7 +44,7 @@ const props = defineProps({
 })
 /* 祝日名の文字の大きさを決める */
 function fontSizeHolodayTitle() {
-  let holidayTitleLength = props.holidayTitle?.length
+  const holidayTitleLength = props.holidayTitle?.length
   if (holidayTitleLength >= 9) {
     return 'font-size-s'
   }
@@ -55,18 +55,9 @@ function fontSizeHolodayTitle() {
 }
 </script>
 
-<style lang="scss" vscoped>
-.date-cell {
-  width: 160px;
-  height: 80px;
-  word-break: break-word;
-}
-.date-holiday-cell {
-  width: 160px;
-  height: 80px;
-  background-color: #f4c9c9;
-}
-.date-cell-margin {
-  margin-top: 5px;
+<style lang="scss" scoped>
+@import '../assets/scss/timetable.scss';
+
+@media only screen and (max-width: 767px) {
 }
 </style>
